// packages/reference-rs/js/tasty/generated/TastyInterfaceSymbol.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TastyJsDoc } from "./TastyJsDoc";
import type { TastyMember } from "./TastyMember";
import type { TastySymbolRef } from "./TastySymbolRef";
import type { TastyTypeParameter } from "./TastyTypeParameter";

export type TastyInterfaceSymbol = { id: string, name: string, library: string, description?: string, descriptionRaw?: string, jsdoc?: TastyJsDoc, typeParameters?: Array<TastyTypeParameter>, members: Array<TastyMember>, extends: Array<TastySymbolRef>, types: Array<TastySymbolRef>, };