// packages/reference-rs/js/tasty/generated/TastyManifest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TastySymbolIndexEntry } from "./TastySymbolIndexEntry";

export type TastyManifest = { version: string, symbolsByName: { [key in string]: string }, symbolsById: { [key in string]: TastySymbolIndexEntry }, };