// packages/reference-rs/js/tasty/generated/TastyTypeAliasSymbol.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TastyJsDoc } from "./TastyJsDoc";
import type { TastyTypeParameter } from "./TastyTypeParameter";
import type { TastyTypeRef } from "./TastyTypeRef";

export type TastyTypeAliasSymbol = { id: string, name: string, library: string, description?: string, descriptionRaw?: string, jsdoc?: TastyJsDoc, typeParameters?: Array<TastyTypeParameter>, definition: TastyTypeRef | null, };